OmniSwitch: Create a shortcut to exposed toggle activity
This exposes the show/hide overlay toggle as a launcher shortcut.
The primary purpose is to expose toggle as an action for our
action handling framework. However, because it is now a shortcut,
it can be accessed by the user anywhere.

NOTE: The service MUST still be turned ON in OmniSwitch settings
for the toggle shortcut to work.

Change-Id: I8fc1896d6d3b51f16e98b307150c9cec7eac24e3
Signed-off-by: Corey Edwards <ensabahnur16@gmail.com>